Move get_query_count to utils.db
